Implemented the feedback by moving the DataGrid grouping example data out of docs-local code and into the shared examples data flow used by docs hosts.
**What Changed**
- Added shared Company model in examples data:
Company.cs
- Added shared Company data provider:
CompanyData.cs
- Added Company API endpoint in docs server:
CompanyController.cs
- Added Company API endpoint in wasm host:
CompanyController.cs
- Refactored multi-level grouping example to load Company data from webapi/company via HttpClient instead of local static records:
DataGridGroupingMultiLevelExample.razor
- Removed Company-specific ad-hoc snippet injection from SectionContent:
SectionContent.razor.cs
- Removed docs-local Company snippet constant:
Snippets.cs

This also removes the need for the previous docs-local Company model path and keeps example data where reviewers requested: shared examples data + host API controllers.
