+++ src/MudBlazor.Examples.Data/CompanyData.cs
@@ -1,34 +1,10 @@
-﻿#nullable enable
-
 using System.Collections.Generic;
+using MudBlazor.Examples.Data.Models;
 
-public class Company
+namespace MudBlazor.Examples.Data;
+
+public static class CompanyData
 {
-    /// <summary>
-    /// Gets or sets the department.
-    /// </summary>
-    public string? Department { get; set; }
-
-    /// <summary>
-    /// Gets or sets the team.
-    /// </summary>
-    public string? Team { get; set; }
-
-    /// <summary>
-    /// Gets or sets the role.
-    /// </summary>
-    public string? Role { get; set; }
-
-    /// <summary>
-    /// Gets or sets the employee name.
-    /// </summary>
-    public string? Name { get; set; }
-
-    /// <summary>
-    /// Gets or sets the score.
-    /// </summary>
-    public int Score { get; set; }
-
     public static IReadOnlyCollection<Company> GetRecords() =>
     [
         new() { Department = "Engineering", Team = "Platform", Role = "Backend", Name = "Alice", Score = 120 },
