+++ src/MudBlazor.Examples.Data/Models/Company.cs
@@ -1,0 +1,29 @@
+namespace MudBlazor.Examples.Data.Models;
+
+public class Company
+{
+    /// <summary>
+    /// Gets or sets the department.
+    /// </summary>
+    public string? Department { get; set; }
+
+    /// <summary>
+    /// Gets or sets the team.
+    /// </summary>
+    public string? Team { get; set; }
+
+    /// <summary>
+    /// Gets or sets the role.
+    /// </summary>
+    public string? Role { get; set; }
+
+    /// <summary>
+    /// Gets or sets the employee name.
+    /// </summary>
+    public string? Name { get; set; }
+
+    /// <summary>
+    /// Gets or sets the score.
+    /// </summary>
+    public int Score { get; set; }
+}
